refactor(catalog): consolidate duplicate catalog helpers into adminCatalogHelpers
