+++ backend/api/src/controllers/admin/catalog/adminBrandController.ts
@@ -7,12 +7,12 @@ import { CATALOG_APPROVAL_STATUS } from "@esparex/contracts";
 import { findBrandByFilter, findCategoryBySlugForCatalog } from '@esparex/core/services/catalog/CatalogBrandModelService';
 import BrandModel from '@esparex/core/models/Brand';
 import CatalogOrchestrator from '@esparex/core/services/catalog/CatalogOrchestrator';
-import { sendCatalogError, QueryRecord, ACTIVE_CATEGORY_QUERY, validateActiveCategories, handleCatalogCreate, handleCatalogUpdate, handleCatalogToggleStatus, handleCatalogReview, sendEmptyPublicList, applyCatalogStatusFilter, hasAdminAccess, CATALOG_PUBLIC_VISIBILITY_QUERY, deriveApprovalStatus } from './shared';
+import { sendCatalogError, QueryRecord, ACTIVE_CATEGORY_QUERY, validateActiveCategories, handleCatalogCreate, handleCatalogUpdate, handleCatalogToggleStatus, handleCatalogReview, sendEmptyPublicList, applyCatalogStatusFilter, hasAdminAccess, CATALOG_PUBLIC_VISIBILITY_QUERY, deriveApprovalStatus } from './adminCatalogHelpers';
 import { logAdminAction } from '../../../utils/adminLogger';
 import { brandCreateSchema, brandUpdateSchema, rejectionSchema } from '@esparex/core/validators/catalog.validator';
 import CategoryQueryBuilder from '@esparex/core/utils/CategoryQueryBuilder';
 import { getCache } from '@esparex/core/utils/redisCache';
-import { catalogCacheKey, applyCacheWriteThrough } from './adminCatalogShared';
+import { catalogCacheKey, applyCacheWriteThrough } from './adminCatalogHelpers';
 import { AppError } from '@esparex/core/utils/AppError';
 
 export const getBrands = async (req: Request, res: Response) => {
