+++ backend/api/src/controllers/admin/catalog/adminCatalogHelpers.ts
@@ -1,21 +1,12 @@
-/**
- * Shared utilities for catalog controllers
- * Extracted from original catalog.content.controller.ts
- *
- * Validation logic is delegated to CatalogValidationService (SSOT).
- * Re-exports keep existing controller imports stable.
- */
-
+import mongoose, { Document, Model as MongooseModel } from 'mongoose';
 import { Request, Response } from 'express';
-import { Document, Model as MongooseModel } from 'mongoose';
 import { z } from 'zod';
 import slugify from 'slugify';
 import { nanoid } from 'nanoid';
+import logger from '@esparex/core/utils/logger';
 import { respond, sendSuccessResponse } from "../../../utils/respond";
 import { sendErrorResponse as sendContractErrorResponse, sendCatalogError } from "../../../utils/errorResponse";
 import { isDuplicateKeyError } from '@esparex/core/utils/errorHelpers';
-
-// Re-export SSOT validation helpers so controllers import from one place.
 import { CATALOG_APPROVAL_STATUS } from '@esparex/contracts';
 import {
     ACTIVE_CATEGORY_QUERY,
@@ -25,9 +16,13 @@ import {
     validateActiveCategories,
     deriveApprovalStatus,
 } from '@esparex/core/services/catalog/CatalogValidationService';
-
 import { logAdminAction } from '../../../utils/adminLogger';
 import { handlePaginatedContent } from "../../../utils/contentHandler";
+import { toOptionalString } from './inputCoercion';
+import { setCache } from '@esparex/core/utils/redisCache';
+import { getVariantsAndModelsForParentModels, getBrandModelsForDuplicateCheck } from '@esparex/core/services/catalog/CatalogBrandModelService';
+import { validateModelHierarchyMutation } from '@esparex/core/services/catalog/CatalogHierarchyService';
+import { detectDuplicateCandidates } from '@esparex/core/services/catalog/CatalogSearchGovernanceService';
 
 export {
     sendCatalogError,
@@ -38,7 +33,8 @@ export {
     CATALOG_PUBLIC_VISIBILITY_QUERY,
     getActiveCategoryIds,
     validateActiveCategories,
-    deriveApprovalStatus
+    deriveApprovalStatus,
+    isDuplicateKeyError
 };
 
 export type CatalogRequest = Request & {
@@ -115,8 +111,6 @@ export const getAdminActorId = (req: Request): string | undefined => {
     return undefined;
 };
 
-// sendCatalogError is imported and re-exported from "../../../utils/errorResponse" (line above)
-
 /**
  * Send Zod validation error response mapping issues to field-level details
  */
@@ -141,9 +135,6 @@ export const sendEmptyPublicList = (res: Response) => {
         }
     }));
 };
-
-// isDuplicateKeyError imported from errorHelpers (SSOT)
-export { isDuplicateKeyError };
 
 /* ======================================================
    GENERIC CATALOG CRUD HANDLERS
@@ -403,7 +394,6 @@ export async function handleCatalogReview<T extends Document>(
             };
         }
 
-
         const item = await model.findByIdAndUpdate(req.params.id, { $set: updates }, { new: true });
         if (!item) {
             return sendContractErrorResponse(req, res, 404, `${model.modelName} not found`);
@@ -420,3 +410,84 @@ export async function handleCatalogReview<T extends Document>(
         return sendCatalogError(req, res, error);
     }
 }
+
+/* ======================================================
+   ADMIN CATALOG SHARED HELPERS
+====================================================== */
+
+export const CATALOG_CACHE_TTL = 300;
+
+export const normalizeCacheValue = (value: unknown): string => {
+    if (Array.isArray(value)) return value.map(normalizeCacheValue).join(',');
+    if (value === undefined || value === null || value === '') return 'all';
+    return encodeURIComponent(String(value));
+};
+
+export const catalogCacheKey = {
+    brands: (categoryId: string) => `catalog:brands:${normalizeCacheValue(categoryId)}`,
+    models: (params: {
+        categoryId?: unknown; brandId?: unknown; parentModelId?: unknown; variantModelId?: unknown;
+        includeVariants?: unknown; treeView?: unknown; search?: unknown; q?: unknown;
+        page?: unknown; limit?: unknown; sort?: unknown;
+    }) => ['catalog:models', `category=${normalizeCacheValue(params.categoryId)}`, `brand=${normalizeCacheValue(params.brandId)}`,
+        `parent=${normalizeCacheValue(params.parentModelId)}`, `variant=${normalizeCacheValue(params.variantModelId)}`,
+        `includeVariants=${normalizeCacheValue(params.includeVariants)}`, `treeView=${normalizeCacheValue(params.treeView)}`,
+        `search=${normalizeCacheValue(params.search ?? params.q)}`, `page=${normalizeCacheValue(params.page ?? 1)}`,
+        `limit=${normalizeCacheValue(params.limit ?? 100)}`, `sort=${normalizeCacheValue(params.sort ?? 'name')}`,
+    ].join(':'),
+};
+
+export const normalizeOptionalObjectIdQuery = (value: unknown): string | undefined => {
+    const raw = Array.isArray(value) ? value[0] : value;
+    if (typeof raw !== 'string') return undefined;
+    const normalized = raw.trim();
+    if (!normalized || normalized === 'all') return undefined;
+    return mongoose.Types.ObjectId.isValid(normalized) ? normalized : undefined;
+};
+
+export const normalizeBooleanQuery = (value: unknown): boolean => {
+    const raw = Array.isArray(value) ? value[0] : value;
+    return raw === true || raw === 'true' || raw === '1';
+};
+
+export const populateModelVariants = async (items: unknown[]) => {
+    const modelIds = items.map((item) => { const m = item as { _id?: unknown; id?: unknown }; return m._id ?? m.id; }).filter(Boolean).map(String);
+    if (modelIds.length === 0) return items;
+    const [variantDocs, variantModelDocs] = await getVariantsAndModelsForParentModels(modelIds);
+    const variantsByModelId = new Map<string, unknown[]>();
+    for (const v of variantDocs) { const mid = String((v as { modelId?: unknown }).modelId ?? ''); if (!mid) continue; const ex = variantsByModelId.get(mid) ?? []; ex.push(v); variantsByModelId.set(mid, ex); }
+    const variantModelsByParentId = new Map<string, unknown[]>();
+    for (const vm of variantModelDocs) { const pid = String((vm as { variantOfModelId?: unknown }).variantOfModelId ?? ''); if (!pid) continue; const ex = variantModelsByParentId.get(pid) ?? []; ex.push(vm); variantModelsByParentId.set(pid, ex); }
+    return items.map((item) => {
+        const plain = typeof (item as { toObject?: () => unknown }).toObject === 'function' ? (item as { toObject: () => Record<string, unknown> }).toObject() : { ...(item as Record<string, unknown>) };
+        const id = String(plain._id ?? plain.id ?? '');
+        return { ...plain, variants: variantsByModelId.get(id) ?? [], variantModels: variantModelsByParentId.get(id) ?? [] };
+    });
+};
+
+export const applyModelHierarchyPayload = async (payload: Record<string, unknown>, options: { existingModel?: unknown } = {}) => {
+    const normalizedPayload = {
+        ...payload,
+        brandId: toOptionalString(payload.brandId) ?? payload.brandId,
+        parentModelId: payload.parentModelId === null ? null : toOptionalString(payload.parentModelId) ?? payload.parentModelId,
+        variantOfModelId: payload.variantOfModelId === null ? null : toOptionalString(payload.variantOfModelId) ?? payload.variantOfModelId,
+    };
+    return validateModelHierarchyMutation(normalizedPayload, { existingModel: options.existingModel as any });
+};
+
+export const logModelDuplicateCandidates = async (req: Request, payload: Record<string, unknown>, options: { excludeId?: string } = {}) => {
+    const name = String(payload.displayName ?? payload.name ?? payload.canonicalName ?? '').trim();
+    const brandId = toOptionalString(payload.brandId);
+    if (!name || !brandId) return;
+    const candidates = await getBrandModelsForDuplicateCheck(brandId, options.excludeId);
+    const dupes = detectDuplicateCandidates(name, candidates as unknown as Record<string, unknown>[]);
+    if (dupes.length > 0) logger.warn('[CatalogSearch] Potential model duplicate candidates detected', { requestPath: req.originalUrl || req.path, candidateCount: dupes.length, input: name, candidates: dupes });
+};
+
+export const applyCacheWriteThrough = (res: Response, cacheKey: string) => {
+    const originalJson = res.json.bind(res);
+    res.json = (body: unknown) => {
+        if (res.statusCode >= 200 && res.statusCode < 300) setCache(cacheKey, body, CATALOG_CACHE_TTL).catch(() => {});
+        return originalJson(body);
+    };
+};
